Started to port some input stuff (created an HAL class)
Created classes for input hal.
Moved the seup code of wii motes into it.
Steal need to add PS3 pad init code.
HAL is a mess because PADs are specific to each console, but the buttons are differents... don't know yet creating a A/Cross button  is better than action button... Wait and see